Move SizeEstimator out of util
