+++ core/src/main/scala/org/apache/spark/storage/MemoryStore.scala
@@ -23,7 +23,8 @@ import java.util.LinkedHashMap
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
-import org.apache.spark.util.{SizeEstimator, Utils}
+import org.apache.spark.SizeEstimator
+import org.apache.spark.util.Utils
 import org.apache.spark.util.collection.SizeTrackingVector
 
 private case class MemoryEntry(value: Any, size: Long, deserialized: Boolean)
