+++ core/src/test/scala/org/apache/spark/util/collection/OpenHashSetSuite.scala
@@ -20,7 +20,7 @@ package org.apache.spark.util.collection
 import org.scalatest.FunSuite
 import org.scalatest.Matchers
 
-import org.apache.spark.util.SizeEstimator
+import org.apache.spark.SizeEstimator
 
 class OpenHashSetSuite extends FunSuite with Matchers {
