+++ core/src/test/scala/org/apache/spark/util/collection/PrimitiveVectorSuite.scala
@@ -19,7 +19,7 @@ package org.apache.spark.util.collection
 
 import org.scalatest.FunSuite
 
-import org.apache.spark.util.SizeEstimator
+import org.apache.spark.SizeEstimator
 
 class PrimitiveVectorSuite extends FunSuite {
