+++ core/src/test/scala/org/apache/spark/util/collection/SizeTrackerSuite.scala
@@ -22,7 +22,7 @@ import scala.util.Random
 
 import org.scalatest.FunSuite
 
-import org.apache.spark.util.SizeEstimator
+import org.apache.spark.SizeEstimator
 
 class SizeTrackerSuite extends FunSuite {
   val NORMAL_ERROR = 0.20
